test(tasks): cover store wake signaling across backends
Move the run-scoped wake assertions into the shared conformance suite so
they run against the sqlite backend too, not just the in-memory store.
The case subscribes to a run and asserts a wake after every run-scoped
mutation, which also exercises sqlite's delegating writes (UpdateRun via
CreateRun, UpdateStep via AppendStep). Backends without the optional
SubscribeRun capability skip rather than fail.
